rea(remove): deletar phone

+++ models/phonebook/remove.js
@@ -1,9 +1,11 @@
 const connection = require('..connection/');
 
-const remove = async (phone_id) => {
-  const query = 'DELETE FROM model_phonebook.phonebook WHERE id = ?';
+const remove = async (phone_id, user_id) => {
+  const query = 'DELETE FROM model_phonebook.phonebook WHERE id = ? AND user_id = ?';
 
-  const [ phonebook ] = await connection.execute(query, [phone_id]);
+  const [ phonebook ] = await connection.execute(query, [phone_id, user_id]);
 
   return phonebook;
 }
+
+module.exports = remove;
